// src/types/Domain.ts
import { z } from "zod";
import { baseEntitySchema } from "./BaseEntity";

export const domainInsertSchema = z.object({
  tenant_id: z.string(),
  domain: z.string(),
  dkim_private_key: z.string().optional(),
  dkim_public_key: z.string().optional(),
  email_api_key: z.string().optional(),
  email_service: z.enum(["mailgun", "mailchannels"]),
});

export const domainSchema = baseEntitySchema.extend(domainInsertSchema.shape);